backend/requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9

# ASGI Server
uvicorn[standard]>=0.25.0
gunicorn>=21.2.0

# Celery & Redis
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# LangChain & LLM
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.6.0

# Auth & Security
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.0

# Utils
python-dateutil>=2.8.2
pytz>=2023.3
